[issue-51] ADMIN: CRUD Sertifikasi Tuk

+++ database/migrations/2020_12_22_180025_create_sertifikasi_tuks_table.php
@@ -17,8 +17,8 @@ class CreateSertifikasiTuksTable extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('sertifikasi_id');
             $table->bigInteger('tuk_id');
-            $table->float('tuk_price_baru');
-            $table->float('tuk_price_perpanjang');
+            $table->decimal('tuk_price_baru', 12);
+            $table->decimal('tuk_price_perpanjang', 12);
             $table->timestamps();
         });
     }
